pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "BDRC_WORK_REPORT_GENERATOR"
version = "0.0.1"
authors = [
  { name="OpenPecha", email="dev@openpecha.org" },
]
description = "A small example package"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "pyewts==0.2.0",
    "PyYAML==6.0",
    "rdflib==6.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pre-commit",
]


[project.urls]
"Homepage" = "https://github.com/OpenPecha/openpecha-project-template"
"Bug Tracker" = "https://github.com/OpenPecha/openpecha-project-template/issues"
